Added storyshots, renamed files
