+++ app/src/ui/components/avatar.stories.js
@@ -11,7 +11,7 @@ import { withKnobs, text, boolean, number } from '@storybook/addon-knobs/react'
 import { Link, Router, browserHistory } from 'react-router-dom'
 import { Grid, Container, Segment } from 'semantic-ui-react'
 
-import Avatar from './Avatar'
+import Avatar from './avatar'
 
 storiesOf('Components', module)
   .addDecorator(StoryRouter())
