+++ app/src/ui/components/avatar.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Avatar from './Avatar';
+import Avatar from './avatar';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
